[REFACTOR] 유저 도메인 CQRS 패턴 적용
