+++ team/src/main/java/sansam/team/user/command/application/dto/JwtToken.java
@@ -1,4 +1,4 @@
-package sansam.team.user.command.dto;
+package sansam.team.user.command.application.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
