+++ team/src/main/java/sansam/team/user/command/application/dto/UserDTO.java
@@ -1,14 +1,10 @@
-package sansam.team.user.command.dto;
+package sansam.team.user.command.application.dto;
 
 import lombok.*;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import sansam.team.user.command.enums.RoleType;
-import sansam.team.user.command.enums.StatusType;
+import sansam.team.user.command.domain.aggregate.RoleType;
+import sansam.team.user.command.domain.aggregate.StatusType;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.Collection;
 
 @Setter
 @Getter
