+++ team/src/main/java/sansam/team/user/command/application/service/UserService.java
@@ -1,20 +1,20 @@
-package sansam.team.user.command.service;
+package sansam.team.user.command.application.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
-import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import sansam.team.common.jwt.JWTUtil;
-import sansam.team.user.command.dto.JwtToken;
-import sansam.team.user.command.dto.LoginRequestDTO;
-import sansam.team.user.command.dto.UserDTO;
-import sansam.team.user.command.dto.UserJoinDTO;
-import sansam.team.user.command.entity.User;
-import sansam.team.user.command.repository.UserRepository;
+import sansam.team.user.command.application.dto.JwtToken;
+import sansam.team.user.command.application.dto.LoginRequestDTO;
+import sansam.team.user.command.application.dto.UserDTO;
+import sansam.team.user.command.application.dto.UserJoinDTO;
+import sansam.team.user.command.domain.aggregate.entity.User;
+import sansam.team.user.command.domain.repository.UserRepository;
+import sansam.team.user.command.infrastructure.repository.JpaUserRepository;
 
 @Service
 @RequiredArgsConstructor
