+++ team/src/main/java/sansam/team/user/command/infrastructure/repository/JpaUserRepository.java
@@ -1,0 +1,12 @@
+package sansam.team.user.command.infrastructure.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import sansam.team.user.command.domain.aggregate.entity.User;
+import sansam.team.user.command.domain.repository.UserRepository;
+
+import java.util.Optional;
+
+public interface JpaUserRepository extends UserRepository, JpaRepository<User, Long> {
+
+
+}
