+++ team/src/test/java/sansam/team/project/query/service/ProjectApplyMemberServiceTest.java
@@ -10,7 +10,7 @@ import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
 import sansam.team.project.query.dto.projectboard.ApplyMemberQueryDTO;
 import sansam.team.project.query.mapper.ProjectApplyMemberMapper;
-import sansam.team.user.command.entity.User;
+import sansam.team.user.command.domain.aggregate.entity.User;
 import sansam.team.project.command.domain.aggregate.ApplyStatus;
 
 import java.time.LocalDateTime;
